Archiving recs
The ability to remove a rec and or a brand from the system without
deleting it (which breaks FKs in logs etc).

Note, if you remove all recs, the brand will still appear on the
directory unless you remove it separately, and vv.

Pass “EXCLUDE”, “ONLY”, or “BOTH” through on the archived parameter for
both ethicalBrands and recommendations get functions